Refactored the RemoteMongoDBService
