+++ packages/realm-web/src/services/index.ts
@@ -1,6 +1,6 @@
 import { Transport } from "../transports";
 
-import { createService as createMongoDBRemoteService } from "./mongodb-remote";
+import { createService as createMongoDBRemoteService } from "./RemoteMongoDBService";
 import { createService as createHttpService } from "./HTTPService";
 
 export function create(transport: Transport): Realm.Services.ServicesFactory {
