--- packages/realm-web/src/services/mongodb-remote/index.ts
@@ -1,4 +1,0 @@
-// @see https://github.com/mongodb/stitch-js-sdk/blob/master/packages/core/services/mongodb-remote/src/internal/CoreRemoteMongoCollectionImpl.ts#L34
-
-export { create as createService } from "./service";
-export { RemoteMongoDBCollection } from "./collection";
